# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Code With Prince",
      "version": "4.11.1",
      "title": "Code With Prince velaiapp"
     
    },
...
      "servers": [
        {
        "url": "/"
        },{
            "url":"/http"
        }
    ],
...
    "components": {
        "schemas": {
            "TodoList": {
                "type": "object",
                "properties": {
                    "number": {
                    "type": "string"
                    }
                   
                }
            },
...
        }
    },
...
    "paths": {
        "/api/sms": {
            "post": {
            "tags": [
                "post  API"
            ],
            "summary": "send otp sms",
            "requestBody": {
            "description": "Post New todo To Database",
            "required": true,
            "content": {
                "application/json": {
                "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                }
            }
    
            }

        },
        "/sms/verification": {
            "post": {
            "tags": [
                "post  API"
            ],
            "summary": "otp verification",
            "requestBody": {
            "description": "Post New todo To Database",
            "required": true,
            "content": {
                "application/json": {
                "schema": {
                    "$ref": "#/components1/schemas/TodoList"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                }
            }}   
        },
...
    }
}